refactor(gfql): chain specializations live next to their admission predicates
The pandas/cuDF lanes (single-node, seeded typed single hop, seeded typed
RETURN-destination) move from chain.py/chain_fast_paths.py into
graphistry/compute/chain_specializations/{admission,hotpaths}.py; the polars
lanes (plain single-hop branches, seeded lane, RETURN-destination) move into
graphistry/compute/gfql/lazy/engine/polars/chain_specializations/. chain.py
and the polars chain only dispatch; chain_fast_paths.py keeps the shared
seed/index helpers. No route admits or declines anything it did not before.
Tests mirror the new module paths.
