+++ graphistry/compute/chain.py
@@ -11,12 +11,7 @@ from graphistry.utils.json import JSONVal
 from .ast import ASTObject, ASTNode, ASTEdge, ASTCall, Direction, from_json as ASTObject_from_json, serialize_binding_ops
 from .typing import DataFrameT, SeriesT
 from .util import generate_safe_column_name
-from .chain_fast_paths import (
-    _seeded_typed_hop_pandas_cudf,
-    _single_node_rows_via_index_or_filter,
-    native_fast_path_admits,
-    _tag_fast_path_aliases,
-)
+from .chain_specializations.hotpaths import _try_chain_fast_path
 from graphistry.compute.validate.validate_schema import validate_chain_schema, validate_graph_shape
 from graphistry.compute.gfql.strictness import StrictInput
 from graphistry.compute.gfql.same_path_types import (
@@ -839,154 +834,6 @@ def _step_with_source_edge_columns(g: Plottable, g_step: Plottable, op: ASTObjec
             or name not in edges.columns or edge_id not in step_edges.columns):
         return g_step
     return g_step.edges(edges[edges[edge_id].isin(step_edges[edge_id])])
-
-
-def _try_chain_fast_path(
-    g_in: Plottable,
-    ops: List[ASTObject],
-    engine_concrete: Engine,
-    start_nodes: Optional[DataFrameT] = None,
-) -> Optional[Plottable]:
-    """Degenerate-shape fast path (pandas/cuDF): node-only ``MATCH (n)`` or a plain
-    single-hop ``MATCH (a)-[e]->(b)`` skip the forward/backward/combine BFS machinery.
-    Returns the result Plottable, or ``None`` to fall through to the full path.
-
-    Same node/edge sets + VALUES as the full machinery (trackA_golden + hop/chain
-    suites); the 1-hop additionally preserves int node dtypes (the full path upcasts
-    int→float via merge — the merge is the artifact, int is the Cypher-conformant type).
-    Gated to unqueried nodes + a plain single-hop edge; NAMED ops are served (the alias
-    flags are reconstructed by `_tag_fast_path_aliases`) except when undirected or when
-    the same alias is reused. filtered-undirected and seeded chains fall through.
-    polars/dask/spark also fall through (own fast path / lazy semantics)."""
-    from graphistry.compute.filter_by_dict import filter_by_dict
-
-    shape = native_fast_path_admits(ops, engine_concrete, start_nodes)
-    if shape is None:
-        return None
-    engine_abs = EngineAbstract(engine_concrete.value)
-
-    def _materialize_fast_path_graph() -> Plottable:
-        from graphistry.compute.ComputeMixin import _coerce_input_formats  # lazy — avoids circular import
-        g = g_in.materialize_nodes(engine=EngineAbstract(engine_concrete.value))
-        return _coerce_input_formats(g, engine_concrete)
-
-    if shape == "single-node":
-        n0 = ops[0]
-        assert isinstance(n0, ASTNode)  # the predicate admitted this shape
-        g = _materialize_fast_path_graph()
-        if g._nodes is None:
-            return None
-        nodes = _single_node_rows_via_index_or_filter(g, n0, engine_abs)
-        if n0._name is not None:
-            alias_was_column = n0._name in nodes.columns
-            if alias_was_column:
-                nodes = nodes.drop(columns=[n0._name])
-            nodes = nodes.assign(**{n0._name: True})
-            other_columns = [c for c in nodes.columns if c != n0._name]
-            if g._node in other_columns:
-                other_columns = [g._node, *[c for c in other_columns if c != g._node]]
-            if alias_was_column:
-                nodes = nodes[[*other_columns, n0._name]]
-            else:
-                nodes = nodes[[*other_columns[:1], n0._name, *other_columns[1:]]]
-            nodes = nodes.reset_index(drop=True)
-        edges = g._edges.iloc[0:0] if g._edges is not None else None
-        return g.nodes(nodes).edges(edges) if edges is not None else g.nodes(nodes)
-
-    n0, e1, n2 = ops
-    assert isinstance(n0, ASTNode) and isinstance(e1, ASTEdge) and isinstance(n2, ASTNode)  # the predicate admitted this shape
-    alias_n0, alias_e1, alias_n2 = n0._name, e1._name, n2._name
-    direction = e1.direction
-    unconstrained = not n0.filter_dict and not n2.filter_dict
-    g = _materialize_fast_path_graph()
-    if g._nodes is None or g._edges is None:
-        return None
-    src, dst, node = g._source, g._destination, g._node
-    if src is None or dst is None or node is None:
-        return None  # no edge/node bindings -> can't fast-path; full path handles it
-    if alias_n0 == node or alias_n2 == node:
-        # A node alias EQUAL TO THE NODE-ID BINDING would make `_tag_fast_path_aliases`
-        # overwrite the id column with the bool flag (destroying the ids) while the full
-        # path raises on the same query ("The column label '<node>' is not unique") —
-        # a wrong-serve found by adversarial parity testing. Decline; never serve.
-        return None
-    if alias_e1 is not None and direction in ("forward", "reverse") \
-            and alias_e1 == (src if direction == "forward" else dst):
-        # An edge alias EQUAL TO THE HOP'S FROM-SIDE BINDING (forward+src / reverse+dst)
-        # made the two lanes return DIFFERENT node sets: the full path's flag overwrite
-        # corrupts its own node reduction, the fast path tags after reducing. TO-side
-        # collisions keep parity (pinned in tests) and stay served. Decline; never serve.
-        return None
-    concat = df_concat(engine_concrete)
-    if unconstrained:
-        # No node filter to reduce by: validate BOTH endpoints against the full
-        # node table (the full path drops dangling edges via its joins). dropna so
-        # a NaN node id can't validate a NaN endpoint — .isin treats NaN as
-        # matchable but the BFS joins never match NaN<->NaN.
-        node_ids = g._nodes[node].dropna()
-        edges = g._edges[g._edges[src].isin(node_ids) & g._edges[dst].isin(node_ids)]
-        if e1.edge_match:
-            # typed edge (e.g. -[:HAS_CREATOR]->) — same edge-frame filter the full
-            # hop applies, so the result set is identical.
-            edges = filter_by_dict(edges, e1.edge_match, engine_abs)
-    else:
-        # #1755 lever-3 seed-first: a seeded 1-hop must be O(result), not O(E).
-        # Reduce edges by the selective node filter(s) BEFORE the typed-edge scan
-        # and endpoint validation, so the expensive object/isin passes run on the
-        # tiny frontier, not all edges. The from-side ids come from the node table
-        # (so that endpoint is validated); the node gather below validates the to
-        # side and drops any edge dangling off the node table.
-        # pandas + cuDF: a scalar-filtered seeded typed hop collapses to a few
-        # DataFrame filters (sub-ms); falls back to the general branch below for
-        # predicates / undirected / missing columns (and non-pandas/cuDF engines).
-        if engine_concrete in (Engine.PANDAS, Engine.CUDF):
-            _fast_res = _seeded_typed_hop_pandas_cudf(g, n0, n2, e1, src, dst, node, direction)
-            if _fast_res is not None:
-                return _tag_fast_path_aliases(
-                    _fast_res, alias_n0, alias_e1, alias_n2, src, dst, node, direction)
-        from_col, to_col = (src, dst) if direction == "forward" else (dst, src)
-        edges = g._edges
-        if n0.filter_dict:
-            from_ids = filter_by_dict(g._nodes, n0.filter_dict, engine_abs)[node]
-            edges = edges[edges[from_col].isin(from_ids)]
-        if e1.edge_match:
-            edges = filter_by_dict(edges, e1.edge_match, engine_abs)
-        if n2.filter_dict:
-            # Apply the destination filter to the SMALL set of gathered dst nodes,
-            # not the full node table — an O(N) object/type scan on all nodes is
-            # exactly the tax we're removing. Gather the frontier's dst nodes
-            # (small isin key), filter those, then drop edges to the losers.
-            to_present = edges[to_col].dropna().unique()
-            to_nodes = filter_by_dict(
-                g._nodes[g._nodes[node].isin(to_present)], n2.filter_dict, engine_abs)
-            edges = edges[edges[to_col].isin(to_nodes[node])]
-        # Validate endpoints + build result nodes on the reduced edge set (small
-        # isin key -> small hashtable; no O(E)-values scan). Engine-agnostic
-        # (pandas + cuDF): gather candidate endpoint nodes, drop edges dangling off
-        # the node table, then keep only nodes still referenced by a surviving edge.
-        ep = concat([
-            edges[[src]].rename(columns={src: node}),
-            edges[[dst]].rename(columns={dst: node}),
-        ]).drop_duplicates()
-        cand = g._nodes[g._nodes[node].isin(ep[node])].drop_duplicates(subset=[node])
-        valid = cand[node].dropna()
-        edges = edges[edges[src].isin(valid) & edges[dst].isin(valid)]
-        final = concat([
-            edges[[src]].rename(columns={src: node}),
-            edges[[dst]].rename(columns={dst: node}),
-        ]).drop_duplicates()
-        nodes = cand[cand[node].isin(final[node])]
-        return _tag_fast_path_aliases(
-            g.nodes(nodes).edges(edges), alias_n0, alias_e1, alias_n2, src, dst, node, direction)
-    endpoints = concat([
-        edges[[src]].rename(columns={src: node}),
-        edges[[dst]].rename(columns={dst: node}),
-    ]).drop_duplicates()
-    nodes = g._nodes[g._nodes[node].isin(endpoints[node])]
-    # match the full path's merge, which collapses duplicate node-id rows
-    nodes = nodes.drop_duplicates(subset=[node])
-    return _tag_fast_path_aliases(
-        g.nodes(nodes).edges(edges), alias_n0, alias_e1, alias_n2, src, dst, node, direction)
 
 
 def reject_alias_named_like_binding(
