+++ graphistry/compute/chain_fast_paths.py
@@ -1,24 +1,17 @@
-"""Seeded typed-hop fast-path specializations for the chain executor.
-
-Extracted verbatim from chain.py (#1755) to keep that orchestrator readable: the seeded
-typed 1-hop pandas/cuDF reduction and the seeded typed RETURN-destination pandas/cuDF and
-polars reductions. Pure code move, no behavior change. chain.py imports from here (one
-direction); this module imports only leaf modules (no back-edge into chain.py).
-
-New fast-path helpers belong HERE, not in chain.py — that is the direction this module was
-created to establish and it is why `_tag_fast_path_aliases` lives alongside the seeded
-reductions rather than next to `_try_chain_fast_path`.
-"""
+"""Seed-resolution and resident-index helpers shared by the chain specializations
+(``chain_specializations/``, ``gfql/lazy/engine/polars/chain_specializations/``) and the Cypher
+lanes in ``gfql_fast_paths.py``. This module imports only leaf modules (no back-edge into
+``chain.py`` or the specialization packages)."""
 # ruff: noqa: E501
 
 from typing import Any, Dict, Literal, Optional, Sequence, Tuple, TYPE_CHECKING, Union, cast
 
 from graphistry.Plottable import Plottable
-from .ast import ASTObject, ASTNode, ASTEdge, Direction
+from .ast import Direction
 from .typing import ArrayLike, ArrayNamespace, DataFrameT, SeriesT
 
 if TYPE_CHECKING:
-    from graphistry.Engine import Engine, EngineAbstract
+    from graphistry.Engine import Engine
     from graphistry.compute.gfql.index.registry import AdjacencyIndex, NodeIdIndex
 
 
@@ -78,45 +71,6 @@ def _tag_fast_path_aliases(
         edges = edges.reset_index(drop=True)
 
     return res.nodes(nodes).edges(edges)
-
-
-NativeFastPathShape = Literal["single-node", "seeded-hop"]
-
-
-def native_fast_path_admits(
-    ops: Sequence[ASTObject], engine: "Engine", start_nodes: Optional[DataFrameT],
-) -> Optional[NativeFastPathShape]:
-    """The shape the pandas/cuDF chain fast path serves for ``ops``, or None when the full
-    path must run. This is the dispatcher's own gate (``chain._try_chain_fast_path`` calls
-    it first), so a test that filters a shape corpus with it exercises exactly what the
-    dispatcher admits: a node-only op without ``query``, or a 3-op plain single hop whose
-    node ops carry no ``query``, whose edge carries no node matches, queries, zero-hop seed
-    or endpoint pruning, whose aliases are distinct, and which is not an undirected hop
-    with names or with node filters. Seeded chains (``start_nodes``) and other engines
-    decline. Frame-dependent conditions (missing frames, alias equal to the node binding)
-    are checked by the body after materialization."""
-    from graphistry.Engine import Engine
-    if engine not in (Engine.PANDAS, Engine.CUDF) or start_nodes is not None:
-        return None
-    if len(ops) == 1:
-        n0 = ops[0]
-        return "single-node" if isinstance(n0, ASTNode) and n0.query is None else None
-    if len(ops) != 3:
-        return None
-    n0, e1, n2 = ops
-    if not (isinstance(n0, ASTNode) and n0.query is None and isinstance(n2, ASTNode) and n2.query is None):
-        return None
-    if not (isinstance(e1, ASTEdge) and e1.is_simple_single_hop()
-            and e1.source_node_match is None and e1.destination_node_match is None
-            and e1.source_node_query is None and e1.destination_node_query is None
-            and e1.edge_query is None and not e1.include_zero_hop_seed and not e1.prune_to_endpoints):
-        return None
-    named = [a for a in (n0._name, e1._name, n2._name) if a is not None]
-    if len(named) != len(set(named)):
-        return None
-    if e1.direction == "undirected" and (n0._name is not None or n2._name is not None or n0.filter_dict or n2.filter_dict):
-        return None
-    return "seeded-hop"
 
 
 def _seeded_scalar_filters(fd: Optional[Dict[str, Any]], df: DataFrameT) -> Optional[Dict[str, Any]]:
@@ -296,6 +250,7 @@ def _seed_rows_via_prop_index_frame(
 
 
 SeedRowsHow = Literal["node_id_index", "property_index", "scan"]
+SeededReturn = Tuple[DataFrameT, DataFrameT, DataFrameT, bool]
 
 
 def _seed_node_rows(
@@ -327,27 +282,6 @@ def _seed_node_rows(
     return _filter_frame(seed, filter_dict if filter_dict is not None else n0f, engine), how
 
 
-def _single_node_rows_via_index_or_filter(
-    g: Plottable, n0: ASTNode, engine_abs: "EngineAbstract",
-) -> DataFrameT:
-    """Resolve a single node op through a resident index or the canonical filter."""
-    from .filter_by_dict import filter_by_dict
-    nodes_df = g._nodes
-    assert nodes_df is not None
-    if not n0.filter_dict:
-        return nodes_df
-    node = g._node
-    n0f = _seeded_scalar_filters(n0.filter_dict, nodes_df) if node is not None else None
-    if node is not None and n0f:
-        nid_ctx = _resident_node_id_index(g, nodes_df, node)
-        rows, how = _seed_node_rows(g, nodes_df, n0f, node, nid_ctx, n0.filter_dict)
-        if how != "scan":
-            _record_native_seed_lane(nodes_df, seam="native_seed_lookup", reason=how, hop_count=0,
-                                     public_seed_scan=node not in n0.filter_dict)
-        return rows
-    return filter_by_dict(nodes_df, n0.filter_dict, engine_abs)
-
-
 def _record_native_seed_lane(
     nodes_df: DataFrameT, *, seam: str, reason: str, hop_count: int, public_seed_scan: bool,
 ) -> None:
@@ -376,330 +310,3 @@ def _index_edge_rows(
         return None
     rows, _ = lookup_edge_rows(adj, arr, xp)
     return take_rows(edges_df, xp.sort(rows) if preserve_input_order else rows, engine)
-
-
-def _indexed_kernel_admits(
-    seed_nodes: DataFrameT, gathered_edges: Optional[DataFrameT], n0f: Dict[str, object],
-    node: str, how: SeedRowsHow, ctx: Tuple["NodeIdIndex", "AdjacencyIndex", ArrayNamespace, "Engine"],
-    n_nodes: int, n_edges: int,
-) -> bool:
-    """Whether the indexed connected-bindings kernel would have served this seeded 1-hop:
-    its seed admission (binding-column integer seed, property-index hit, or a scan on a
-    graph with fewer nodes than edges) and its frontier and gather cost gates."""
-    from numbers import Integral
-    from graphistry.compute.gfql.index.cost import cost_gate_frac
-    _, adj, _, engine = ctx
-    seed_val = n0f.get(node)
-    seeded_on_binding = isinstance(seed_val, Integral) and not isinstance(seed_val, bool)
-    if not (seeded_on_binding or how == "property_index" or n_nodes < n_edges):
-        return False
-    frac = cost_gate_frac(engine)
-    n_frontier = int(seed_nodes[node].nunique()) if not hasattr(seed_nodes, "get_column") \
-        else int(seed_nodes.get_column(node).n_unique())
-    if n_frontier >= frac * adj.n_keys:
-        return False
-    return gathered_edges is not None and len(gathered_edges) < frac * n_edges
-
-
-def _seeded_typed_hop_pandas_cudf(
-    g: Plottable, n0: ASTNode, n2: ASTNode, e1: ASTEdge,
-    src: str, dst: str, node: str, direction: Direction,
-) -> Optional[Plottable]:
-    """#1755 lever-3: engine-generic (pandas + cuDF) fast path for a scalar-filtered
-    seeded typed 1-hop. Value-identical to the general seeded branch for the covered
-    shape (all node/edge filters are plain scalars, directed) — same rows, columns,
-    and dtypes; row order and RangeIndex may differ — collapsing it into a
-    few DataFrame filters so a seeded lookup lands sub-ms. Uses only the shared
-    pandas/cuDF DataFrame API (no numpy array drops) so the same body runs on both
-    engines. Returns None to fall back for anything it does not cover (predicates,
-    undirected, missing columns) — the caller then runs the general branch."""
-    if direction == "undirected":
-        return None
-
-    nodes_df, edges_df = g._nodes, g._edges
-    if nodes_df is None or edges_df is None:
-        return None
-    n0f = _seeded_scalar_filters(n0.filter_dict, nodes_df)
-    n2f = _seeded_scalar_filters(n2.filter_dict, nodes_df)
-    ef = _seeded_scalar_filters(e1.edge_match, edges_df)
-    if n0f is None or n2f is None or ef is None:
-        return None
-    from_col, to_col = (src, dst) if direction == "forward" else (dst, src)
-
-    # from-side seed FIRST: reduce edges to the seed's out-edges before the
-    # edge_match compare, so the type filter runs on the tiny frontier rather than
-    # all edges — this is what makes a seeded lookup sub-ms. The id filter goes
-    # first (int, unique -> ~1 row in one pass) so any remaining object filters
-    # (label__X->type) run on that tiny survivor frame, not the whole node table.
-    # Resident-index acceleration (#1658 x #1755): when node-id + directional
-    # adjacency indexes are valid for these exact frames, the O(N) seed scan, the
-    # O(E) frontier isin, and the O(N) candidate gather become positional lookups.
-    # Any decline (no index, stale fingerprint, unsafe id cast) falls back to the
-    # scan body below — identical results either way, only speed differs.
-    # Decoupled index use: the seed row lookup uses the node-id index ONLY when
-    # the seed filter includes the binding column, but the frontier edge gather
-    # (CSR adjacency) and candidate gathers (node-id index) engage regardless of
-    # HOW the seed rows were found — their inputs are binding-column values, which
-    # are the index key domain. (LDBC/user pattern: seed on the `id` PROPERTY
-    # while the graph binds a different key column — previously disqualified the
-    # whole index path.)
-    ctx = _resident_seed_indexes(g, nodes_df, edges_df, node, src, dst, direction) if n0f else None
-    seed_nodes = edges = cand = None
-    if ctx is not None:
-        nid, adj, xp, idx_engine = ctx
-        seed_nodes, _ = _seed_node_rows(g, nodes_df, n0f, node, (nid, xp, idx_engine), n0.filter_dict)
-        edges = _index_edge_rows(adj, seed_nodes[node], xp, idx_engine, edges_df)
-        if edges is not None:
-            if ef:
-                for k, v in ef.items():
-                    edges = edges[edges[k] == v]
-            if 'cudf' in str(type(edges).__module__):
-                import cudf as _cd  # type: ignore
-                endpoint_ids = _cd.concat([edges[src], edges[dst]])
-            else:
-                import pandas as _pd
-                endpoint_ids = _pd.concat([edges[src], edges[dst]])
-            cand = _index_node_rows(nid, endpoint_ids, xp, idx_engine, nodes_df)
-    served_via_index = cand is not None
-    if cand is None:
-        if n0f:
-            seed_nodes = nodes_df
-            for k, v in sorted(n0f.items(), key=lambda kv: 0 if kv[0] == node else 1):
-                seed_nodes = seed_nodes[seed_nodes[k] == v]
-            edges = edges_df[edges_df[from_col].isin(seed_nodes[node].dropna())]
-        else:
-            edges = edges_df
-        if ef:  # typed edge (edge_match) — now on the reduced frontier
-            for k, v in ef.items():
-                edges = edges[edges[k] == v]
-
-        # Gather candidate endpoint nodes (both endpoints of surviving edges), then run
-        # the dest filter, dangling-edge drop and final-node selection on the small
-        # candidate/edge frames. Selecting from nodes_df keeps only real nodes, so the
-        # endpoint-in-nodes check subsumes the old NaN-endpoint guard. Membership sets
-        # are dropna()'d: pandas .isin matches NaN<->NaN, but the general branch's BFS
-        # joins never join on null keys, so a null id/endpoint must not link.
-        cand = nodes_df[
-            nodes_df[node].isin(edges[src].dropna()) | nodes_df[node].isin(edges[dst].dropna())
-        ].drop_duplicates(subset=[node])
-    assert edges is not None and cand is not None  # both branches above assign
-    if served_via_index:
-        _record_native_seed_lane(nodes_df, seam="native_seeded_hop", reason="served", hop_count=1,
-                                 public_seed_scan=node not in n0f)
-    if n2f:  # destination-node filter (to-side)
-        n2_cand = cand
-        for k, v in n2f.items():
-            n2_cand = n2_cand[n2_cand[k] == v]
-        n2_ok = n2_cand[node]
-    else:
-        n2_ok = cand[node]
-    to_vals = edges[to_col]
-    keep = edges[src].isin(cand[node].dropna()) & edges[dst].isin(cand[node].dropna()) & to_vals.isin(n2_ok.dropna())
-    edges = edges[keep]
-    cand = cand[cand[node].isin(edges[src]) | cand[node].isin(edges[dst])]
-    return g.nodes(cand).edges(edges)
-
-
-SeededReturn = Tuple[DataFrameT, DataFrameT, DataFrameT, bool]
-
-
-def _seeded_typed_return_dst_pandas_cudf(
-    g: Plottable, n0: ASTNode, n2: ASTNode, e1: ASTEdge,
-    src: str, dst: str, node: str, direction: Direction,
-) -> Optional[SeededReturn]:
-    """#1755 cypher RETURN-alias fast path: like _seeded_typed_hop_pandas_cudf but
-    returns ONLY the destination (RETURN-alias) node rows + surviving edges — no
-    seed-node gather, no Plottable round-trip — so the seeded cypher projection
-    lands sub-ms. Engine-generic (pandas + cuDF): only the shared DataFrame API,
-    no numpy array drops. Returns ``(dst_node_rows, edges)`` or None to fall back."""
-    if direction == "undirected":
-        return None
-    nodes_df, edges_df = g._nodes, g._edges
-    if nodes_df is None or edges_df is None:
-        return None
-    n0f = _seeded_scalar_filters(n0.filter_dict, nodes_df)
-    n2f = _seeded_scalar_filters(n2.filter_dict, nodes_df)
-    ef = _seeded_scalar_filters(e1.edge_match, edges_df)
-    if n0f is None or n2f is None or ef is None or not n0f:
-        return None
-    from_col, to_col = (src, dst) if direction == "forward" else (dst, src)
-    # id-first seed reduction: filter by the id column first (int/unique -> ~1 row)
-    # so any remaining object filters (label__X->type) run on the tiny survivor
-    # frame, never materializing an object column over the whole node table.
-    # Membership sets are dropna()'d: pandas .isin matches NaN<->NaN, but the full
-    # pipeline's joins never join on null keys, so a null id/endpoint must not link.
-    ctx = _resident_seed_indexes(g, nodes_df, edges_df, node, src, dst, direction)
-    nid_ctx = (ctx[0], ctx[2], ctx[3]) if ctx is not None else _resident_node_id_index(g, nodes_df, node)
-    seed_nodes, how = _seed_node_rows(g, nodes_df, n0f, node, nid_ctx, n0.filter_dict)
-    edges = dstn = None
-    kernel_admits = False
-    if ctx is not None:
-        nid, adj, xp, idx_engine = ctx
-        edges = _index_edge_rows(adj, seed_nodes[node], xp, idx_engine, edges_df)
-        kernel_admits = _indexed_kernel_admits(
-            seed_nodes, edges, n0f, node, how, ctx, len(nodes_df), len(edges_df))
-        if edges is not None:
-            if ef:
-                for k, v in ef.items():
-                    edges = edges[edges[k] == v]
-            dstn = _index_node_rows(nid, edges[to_col], xp, idx_engine, nodes_df)
-    if dstn is None:
-        edges = edges_df[edges_df[from_col].isin(seed_nodes[node].dropna())]
-        if ef:
-            for k, v in ef.items():
-                edges = edges[edges[k] == v]
-        # destination nodes = real nodes that are edge to-endpoints, then the dest
-        # filter, dangling-edge drop and dedup on the small dst/edge frames.
-        dstn = nodes_df[nodes_df[node].isin(edges[to_col].dropna())]
-    assert edges is not None and dstn is not None  # both branches above assign
-    if n2f:
-        for k, v in n2f.items():
-            dstn = dstn[dstn[k] == v]
-    edges = edges[edges[to_col].isin(dstn[node].dropna())]
-    dstn = dstn[dstn[node].isin(edges[to_col].dropna())].drop_duplicates(subset=[node])
-    return dstn, edges, seed_nodes, kernel_admits
-
-
-def _seeded_typed_return_dst_polars(
-    g: Plottable, n0: ASTNode, n2: ASTNode, e1: ASTEdge,
-    src: str, dst: str, node: str, direction: Direction,
-    preserve_input_order: bool = False,
-    index_ctx: Optional[Tuple["NodeIdIndex", "AdjacencyIndex", ArrayNamespace, "Engine"]] = None,
-) -> Optional[SeededReturn]:
-    """#1755 polars analog of _seeded_typed_return_dst_pandas_cudf: same seed-first
-    reduction (seed out-edges -> typed-edge filter -> destination nodes) expressed
-    with polars filters, so a seeded cypher RETURN on polars/polars-gpu also lands
-    sub-ms. Returns ``(dst_node_rows, edges)`` (polars frames) or None to fall back
-    to the full lazy pipeline. Value-identical node set to the full path for the
-    covered shape (scalar filters, directed, single hop); row order may differ."""
-    import polars as pl
-    from graphistry.compute.gfql.lazy.engine.polars.predicates import filter_by_dict_polars
-    if direction == "undirected":
-        return None
-    nodes_df, edges_df = g._nodes, g._edges
-    # Eager polars frames only: LazyFrame has no get_column, and mixed-engine
-    # node/edge frames must take the full path — decline rather than crash.
-    if not isinstance(nodes_df, pl.DataFrame) or not isinstance(edges_df, pl.DataFrame):
-        return None
-
-    n0f = _seeded_scalar_filters(n0.filter_dict, nodes_df)
-    n2f = _seeded_scalar_filters(n2.filter_dict, nodes_df)
-    ef = _seeded_scalar_filters(e1.edge_match, edges_df)
-    if n0f is None or n2f is None or ef is None or not n0f:
-        return None
-    from_col, to_col = (src, dst) if direction == "forward" else (dst, src)
-
-    # from-side seed: reduce the node frame to the seed rows, take their ids.
-    # Membership sets are drop_nulls()'d (null ids/endpoints never link, matching
-    # the full pipeline's joins) and passed via .implode() (Series-arg is_in is
-    # deprecated in polars 1.42, see polars#22149).
-    ctx = index_ctx if index_ctx is not None else _resident_seed_indexes(
-        g, nodes_df, edges_df, node, src, dst, direction)
-    nid_ctx = (ctx[0], ctx[2], ctx[3]) if ctx is not None else _resident_node_id_index(g, nodes_df, node)
-    seed_nodes, how = _seed_node_rows(g, nodes_df, n0f, node, nid_ctx, n0.filter_dict)
-    edges = dstn = None
-    kernel_admits = False
-    if ctx is not None:
-        nid, adj, xp, idx_engine = ctx
-        edges = _index_edge_rows(
-            adj, seed_nodes.get_column(node), xp, idx_engine, edges_df,
-            preserve_input_order=preserve_input_order)
-        kernel_admits = _indexed_kernel_admits(
-            seed_nodes, edges, n0f, node, how, ctx, len(nodes_df), len(edges_df))
-        if edges is not None:
-            edges = filter_by_dict_polars(edges, e1.edge_match)
-            dstn = _index_node_rows(nid, edges.get_column(to_col), xp, idx_engine, nodes_df)
-    if dstn is None:
-        from_ids = seed_nodes.get_column(node).drop_nulls()
-        if from_ids.len() == 0:
-            return nodes_df.clear(), edges_df.clear(), seed_nodes, kernel_admits
-        edges = edges_df.filter(pl.col(from_col).is_in(from_ids.implode()))
-        edges = filter_by_dict_polars(edges, e1.edge_match)
-        dst_ids = edges.get_column(to_col).drop_nulls().unique()
-        dstn = nodes_df.filter(pl.col(node).is_in(dst_ids.implode()))
-    assert edges is not None and dstn is not None  # both branches above assign
-    dstn = filter_by_dict_polars(dstn, n2.filter_dict)
-    # drop dangling edges + dedup destination nodes (mirror the pandas tail)
-    keep_ids = dstn.get_column(node).drop_nulls()
-    edges = edges.filter(pl.col(to_col).is_in(keep_ids.implode()))
-    dstn = dstn.filter(pl.col(node).is_in(edges.get_column(to_col).implode())).unique(subset=[node], maintain_order=True)
-    return dstn, edges, seed_nodes, kernel_admits
-
-
-def polars_seeded_lane_admits(ops: Sequence[ASTObject]) -> bool:
-    """Whether the polars seeded lane's shape gate admits ``ops``: a 3-op directed simple
-    single hop whose seed node carries a filter, with no node queries, endpoint matches,
-    endpoint or edge queries, zero-hop seed or endpoint pruning. The dispatcher calls this
-    first; the frame conditions (polars frames, matching id dtypes, valid resident indexes,
-    scalar-only filters, no colliding aliases) are decided by the body and can still decline
-    an admitted shape."""
-    if len(ops) != 3:
-        return False
-    n0, e1, n2 = ops
-    if not (isinstance(n0, ASTNode) and isinstance(n2, ASTNode) and isinstance(e1, ASTEdge)):
-        return False
-    return not (n0.query is not None or n2.query is not None or not n0.filter_dict
-                or not e1.is_simple_single_hop() or e1.direction not in ("forward", "reverse")
-                or e1.source_node_match is not None or e1.destination_node_match is not None
-                or e1.source_node_query is not None or e1.destination_node_query is not None
-                or e1.edge_query is not None or e1.include_zero_hop_seed or e1.prune_to_endpoints)
-
-
-def _try_seeded_chain_polars(g: Plottable, ops: Sequence[ASTObject]) -> Optional[Plottable]:
-    """Serve a native directed scalar hop through the resident seed indexes, preserving
-    Polars table order and aliases; declines (None) without valid resident indexes."""
-    import polars as pl
-    from graphistry.compute.gfql.index.api import _record_indexed_traversal
-    if not polars_seeded_lane_admits(ops):
-        return None
-    n0, e1, n2 = ops
-    assert isinstance(n0, ASTNode) and isinstance(e1, ASTEdge) and isinstance(n2, ASTNode) and n0.filter_dict  # the predicate admitted this shape
-    nodes, edges = g._nodes, g._edges
-    node, src, dst = g._node, g._source, g._destination
-    if (not isinstance(nodes, pl.DataFrame) or not isinstance(edges, pl.DataFrame)
-            or node is None or src is None or dst is None):
-        return None
-    if nodes.schema[node] != edges.schema[src] or nodes.schema[node] != edges.schema[dst]:
-        return None
-    aliases = [op._name for op in ops if op._name is not None]
-    if len(aliases) != len(set(aliases)):
-        return None
-    if any(name in nodes.columns for name in (n0._name, n2._name) if name is not None):
-        return None
-    if e1._name is not None and e1._name in edges.columns:
-        return None
-    ctx = _resident_seed_indexes(g, nodes, edges, node, src, dst, e1.direction)
-    if ctx is None:
-        return None
-    reduced = _seeded_typed_return_dst_polars(
-        g, n0, n2, e1, src, dst, node, e1.direction, preserve_input_order=True, index_ctx=ctx)
-    if reduced is None:
-        return None
-    _, kept_edges, _, _ = reduced
-    if not isinstance(kept_edges, pl.DataFrame):
-        return None
-    endpoint_ids = pl.concat([kept_edges.get_column(src), kept_edges.get_column(dst)]).drop_nulls().unique()
-    nid_ctx = _resident_node_id_index(g, nodes, node)
-    result_nodes = None
-    if nid_ctx is not None:
-        nid, xp, engine = nid_ctx
-        result_nodes = _index_node_rows(nid, endpoint_ids, xp, engine, nodes, preserve_input_order=True)
-    if result_nodes is None:
-        result_nodes = nodes.filter(pl.col(node).is_in(endpoint_ids.implode()))
-        if result_nodes.get_column(node).n_unique() != result_nodes.height:
-            return None
-    if not isinstance(result_nodes, pl.DataFrame):
-        return None
-    from_col, to_col = (src, dst) if e1.direction == "forward" else (dst, src)
-    flags = [
-        pl.col(node).is_in(kept_edges.get_column(endpoint).implode()).fill_null(False).alias(name)
-        for name, endpoint in ((n0._name, from_col), (n2._name, to_col)) if name is not None
-    ]
-    if flags:
-        result_nodes = result_nodes.with_columns(flags)
-    if e1._name is not None:
-        kept_edges = kept_edges.with_columns(pl.lit(True).alias(e1._name))
-    _record_indexed_traversal(
-        seam="native_seeded_hop", engine=ctx[3], served=True, reason="served", hop_count=1,
-        public_seed_scan=node not in n0.filter_dict, hop_details=[{"hop": 1}])
-    return g.nodes(result_nodes).edges(kept_edges)
