+++ graphistry/compute/gfql/lazy/engine/polars/chain_specializations/__init__.py
@@ -1,0 +1,6 @@
+"""Chain specializations for the polars engine: admission predicates and their lanes."""
+from .admission import PolarsPlainSingleHopShape, polars_plain_single_hop_admits, polars_seeded_lane_admits
+from .hotpaths import _plain_seeded_index_hop_polars, _plain_single_hop_polars, _try_seeded_chain_polars
+
+__all__ = ["PolarsPlainSingleHopShape", "polars_plain_single_hop_admits", "polars_seeded_lane_admits",
+           "_plain_seeded_index_hop_polars", "_plain_single_hop_polars", "_try_seeded_chain_polars"]
